Remove incorrect .real casts from slogdet sign computations
For complex matrices jnp.sign returns a unit complex number (the phase
of the determinant), matching numpy.linalg.slogdet convention. The
.real / .astype(real_dtype) casts were silently discarding that phase.

Cholesky is unaffected (sign is always ±1 from a static Python int).
HEVD is unaffected (Hermitian eigenvalues are always real).
All other solvers (LU, QR, Diagonal, Triangular, Tridiagonal) now return
the natural dtype of jnp.sign, preserving the complex phase.

+++ lineax/_solver/lu.py
@@ -101,7 +101,7 @@ class LU(AbstractDirectLinearSolver[_LUState]):
         num_swaps = jnp.sum(piv != jnp.arange(n, dtype=piv.dtype))
         perm_sign = jnp.where(num_swaps % 2 == 0, 1, -1)
         diag_sign = jnp.prod(jnp.sign(jnp.diag(lu)))
-        sign = (perm_sign * diag_sign).astype(lu.real.dtype)
+        sign = perm_sign * diag_sign
         lad = jnp.sum(jnp.log(jnp.abs(jnp.diag(lu))))
         return sign, lad
